vacancies/pyproject.toml
[project]
name = "vacancies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloud-sql-python-connector[asyncpg]>=1.16.0",
    "langchain>=0.3.14",
    "langchain-google-vertexai>=2.0.11",
    "playwright>=1.49.1",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.7.1",
    "sqlalchemy>=2.0.37",
    "google-cloud-logging>=3.11.3",
    "pg8000>=1.31.2",
    "opentelemetry-sdk>=1.31.1",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0",
    "picologging>=0.9.3",
    "python-json-logger>=3.3.0",
]

[tool.uv]
dev-dependencies = [
    "debugpy>=1.8.11",
    "ipykernel>=6.29.5",
    "pip>=24.3.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.1",
]

[tool.ruff]
extend = "../ruff.toml"

[tool.setuptools]
package-dir = {"" = "app"}
packages = ["vacancies"]
